Move Status class to RoutingPolicy
